requirements.txt
pandas
requests
boto3
snowflake-connector-python
python-dotenv